refactor: separated transaction, input, output, comparators

+++ src/s2f/architecture/blockchain/pow/data/UtxoSet.h
@@ -35,7 +35,7 @@ namespace s2f::chain::pow
          * @param vout Transaction outpoint index.
          * @return The amount in the outpoint or -1.
          */
-        uint32_t getCoin(const sha256digest &txid, int vout) const;
+        uint64_t getCoin(const sha256digest &txid, int vout) const;
 
         /**
          * Invalidates the coin value at txid:vout so that it cannot be spent
